Apply_SkillDescriptorInfoクラスの分離。説明文もローカライズ
